+++ requirements.txt
@@ -1,1 +1,4 @@
 httpx
+click
+flask
+flask_cors
